fix: expose as boolean value.
